Particle selection added -> uploading and deletion may need some changes
